customers/elkj%C3%B8p/div4c/shield misc
More minor cleanups, fixes a few outstanding warnings.

Make a couple of wrapper files and use funky pragmas to work disable warnings on acouple of files. Convert a couple of ints to size_t and make a couple of conversions in the parser explicit. No warnings with Wall remain, though a couple remain with -pedantic.

git-svn-id: https://projects.freecode.no/internal/svn/customers/elkj%C3%B8p/div4c@1561 3f367b02-151f-0410-a67f-93a54fc96dc6

+++ shield/trunk/src/transform_cast.cc
@@ -67,7 +67,7 @@ namespace shield
 	}
       ;
 
-      for (int i=0; i<(sizeof(order)/sizeof(data_type)); i++)
+      for (size_t i=0; i<(sizeof(order)/sizeof(data_type)); i++)
 	{
 	  if (desired & order[i])
 	    {
